first attempt at splitting LocationPickerVidewModel into two sourceSets
